The additions made with this commit are as follows:
-Changed classifiers of isStateView() global methods in the
GlobalButtonVars class to static so that they could be used globally.

-Modified the EventEditor class so that the buttons for selecting
whether or not to create the event as team or personal depends
(as a default) on the view that the user has previously selected.
If the view is both, then it is assumed that the default classification
of the event should be personal.

-Added the database linkage to the "View Both Calendars" button so that
the content of the events that are stored in the local EventModel class
now appear in the CalendarSidebar view when the refresh button is
pressed by the user.

-Fixed the problem where end dates submitted during event creation were
not appearing in the CalendarSidebar view.

-Wrote a complete suite of test cases for the CategoryEntityManager
class. Eclmma now reports 93% coverage. This is the best that can
be done because exceptions within the makeEntity, update, getEntity,
and deleteEntity rely on other methods that would need to be broken
in order for those exceptions to be thrown.